+++ packages/zod/src/index.ts
@@ -5,6 +5,7 @@ import {
   deepMerge,
   ensureInterfaceType,
   mapValue,
+  weave,
   weaverContext,
 } from "@gqloom/core"
 import { LoomObjectType } from "@gqloom/core"
@@ -87,6 +88,15 @@ export class ZodWeaver {
           }
         : ZodWeaver.getGraphQLTypeBySelf,
     })
+  }
+
+  /**
+   * Weave a GraphQL Schema from resolvers with zod schema
+   * @param inputs Resolvers, Global Middlewares, WeaverConfigs Or SchemaWeaver
+   * @returns GraphQL Schema
+   */
+  static weave(...inputs: Parameters<typeof weave>) {
+    return weave(ZodWeaver, ...inputs)
   }
 
   static toNullableGraphQLType(schema: Schema): GraphQLOutputType {