+++ packages/federation/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## next (YYYY-MM-DD)
 
+* Refactor: rename `FederatedSchemaWeaver` to `FederatedSchemaLoom`
+
 ## 0.4.0 (2024-11-12)
 
 * Refactor: follow standard-schema ([#7](https://github.com/modevol-com/gqloom/pull/7))